requirements.txt
# Core dependencies
numpy>=1.19.0
gym>=0.17.3
tensorflow>=2.4.0

# Configuration
PyYAML>=5.4.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Logging and monitoring
tensorboard>=2.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Utilities
tqdm>=4.60.0

# Optional: for notebook support
jupyter>=1.0.0
ipykernel>=5.5.0

# Optional: for advanced visualization
plotly>=5.0.0

